# specs/openapi/dist/api-bundle.yaml
openapi: 3.1.0
info:
  title: Virtual Companion API
  version: 0.1.0
  description: Contract-first HTTP API (ADR-0002). This document is the single source of truth; Java interfaces and TypeScript
    clients are generated from it and must not be hand-edited or used to back-drive the contract. The ErrorEnvelope shape
    is uniform across every error response; NOT_FOUND_OR_FORBIDDEN hides resource existence so a missing resource and a forbidden
    one are indistinguishable to the caller.
security:
- bearerAuth: []
paths:
...
  /api/v1/conversations/{conversationId}/generations:
    post:
      operationId: sendGeneration
      summary: Idempotently send a Chat turn (create or rejoin a generation)
      description: Creates a new generation for the conversation or rejoins the existing one for the supplied idempotency
        key (vc.receive_generation). The caller identity is the owner; owner_user_id is never a request field. Requires a
        server-verified authentication context. Cross-conversation or cross-owner references never disclose existence.
      parameters:
      - $ref: '#/components/parameters/ConversationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendGenerationRequest'
      responses:
        '200':
          description: The generation (newly created or rejoined)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Generation'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource absent or not owned by the caller; existence is not disclosed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          description: RATE_LIMITED — shared per-owner generation admission rejected the request
          headers:
            Retry-After:
              description: Seconds until the caller should retry
              schema:
                type: integer
                minimum: 1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
  /api/v1/generations/{generationId}/snapshot:
    get:
      operationId: getGenerationSnapshot
      summary: Read a consistent generation snapshot
      description: Returns the generation status, the final assistant message id (if any), and the ordered realtime event
        stream up to a consistent point (vc.read_generation_snapshot). A foreign or absent id never discloses existence. Requires
        a server-verified authentication context.
      parameters:
      - $ref: '#/components/parameters/GenerationId'
      responses:
        '200':
          description: A consistent generation snapshot
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GenerationSnapshot'
        '401':
          description: No valid server-verified authentication context
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '404':
          description: Resource absent or not owned by the caller; existence is not disclosed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorEnvelope:
      type: object
      description: Unified error shape. NOT_FOUND_OR_FORBIDDEN never discloses whether a resource exists; every error response
        uses this envelope.
      required:
      - code
      - message
      properties:
        code:
          $ref: '#/components/schemas/ErrorCode'
        message:
          type: string
          description: Stable, non-sensitive human-readable explanation
        details:
          type: object
          additionalProperties: true
          description: Optional structured context (never sensitive data)
    ErrorCode:
      type: string
      description: Stable error code catalog (specs/catalog/error-codes.yaml)
      enum:
      - AUTHENTICATION_REQUIRED
      - ACCESS_DENIED
      - NOT_FOUND_OR_FORBIDDEN
      - AGE_VERIFICATION_REQUIRED
      - MODEL_NO_ELIGIBLE_DEPLOYMENT
      - MODEL_ATTEMPT_TIMEOUT
      - MODEL_ALL_UNAVAILABLE
      - GENERATION_CANCELLED
      - GENERATION_OUTPUT_BLOCKED
      - STREAM_GAP_EXPIRED
      - STREAM_RESET_REQUIRED
      - MEMORY_CONFIRMATION_REQUIRED
      - MEMORY_FORBIDDEN_CONTENT
      - CATALOG_DRIFT_DETECTED
      - BETA_OPERATIONS_NOT_READY
      - INVALID_REQUEST
      - AUTH_RATE_LIMITED
...
    Generation:
      type: object
      description: A generation owned by exactly one user within a conversation. One logical request has one stable logicalGenerationId
        (INV-GEN-001); status follows the generation-states catalog; mode is the interaction mode captured at reception (CHAT-MODE).
      required:
      - generationId
      - conversationId
      - logicalGenerationId
      - status
      - mode
      properties:
        generationId:
          type: string
          description: Stable owner-scoped generation identifier
        conversationId:
          type: string
          description: Owning conversation identifier
        logicalGenerationId:
          type: string
          description: Idempotent logical label shared by create and rejoin
        status:
          type: string
          description: Current generation-states catalog code
        mode:
          $ref: '#/components/schemas/InteractionModeCode'
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp (RFC 3339)
    GenerationSnapshot:
      type: object
      description: 'A consistent point-in-time read of a generation: status, final assistant message id and the ordered realtime
        event stream (INV-RT-001). usage is present only after finalize settles the provider usage (USAGE-VIZ).'
      required:
      - status
      - events
      properties:
        status:
          type: string
          description: Current generation-states catalog code
        assistantMessageId:
          type: string
          description: Final assistant message id once committed, absent otherwise
        events:
          type: array
          description: Ordered realtime events up to a consistent point
          items:
            type: object
            additionalProperties: true
        usage:
          $ref: '#/components/schemas/GenerationUsage'
    GenerationUsage:
      type: object
      description: Settled provider token usage for one generation (USAGE-VIZ).
      required:
      - inputTokens
      - outputTokens
      properties:
        inputTokens:
          type: integer
          format: int64
          minimum: 0
          description: Provider input (prompt) tokens settled at finalize
        outputTokens:
          type: integer
          format: int64
          minimum: 0
          description: Provider output (completion) tokens settled at finalize
...
    SendGenerationRequest:
      type: object
      description: Body for idempotently sending a Chat turn; mode selects the interaction mode for the turn (CHAT-MODE).
      required:
      - idempotencyKey
      properties:
        idempotencyKey:
          type: string
          description: Caller-supplied idempotency label; reuse to rejoin the same generation
        userContent:
          type: string
          description: The user's chat content for this turn
        sourceUserMessageId:
          type: string
          description: Existing user message to regenerate against (GEN-VER / FR-CHAT-003). When set, the server does not
            insert a second user row and the new generation becomes the selected version.
        mode:
          $ref: '#/components/schemas/InteractionModeCode'
    InteractionModeCode:
      type: string
      description: 'Turn-level interaction mode captured at reception (CHAT-MODE). AUTO lets the persona''s default mode rule
        (gentle-listener: LISTEN); LISTEN asks the companion to listen and reflect without unsolicited advice; DISCUSS opens
        an active exchange with questions and analysis; CASUAL is a light everyday stance (FR-CHAT-002). Rejoining the same
        generation via idempotencyKey keeps the first reception''s mode.'
      enum:
      - AUTO
      - LISTEN
      - DISCUSS
      - CASUAL
      default: AUTO
...
  parameters:
...
    ConversationId:
      name: conversationId
      in: path
      required: true
      schema:
        type: string
      description: Owner-scoped conversation identifier
    GenerationId:
      name: generationId
      in: path
      required: true
      schema:
        type: string
      description: Owner-scoped generation identifier
...